remove example adapters
