+++ .changeset/healthy-hornets-compete.md
@@ -2,4 +2,4 @@
 '@chainlink/ea-scripts': minor
 ---
 
-Added new scripts to generate v3 adapters
+Changed EA generation script
