--- packages/examples/composite/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "@chainlink/example-composite-adapter",
-  "version": "1.3.20",
-  "description": "Chainlink example composite adapter.",
-  "keywords": [
-    "Chainlink",
-    "LINK",
-    "BTC",
-    "Bitcoin",
-    "blockchain",
-    "oracle"
-  ],
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "repository": {
-    "url": "https://github.com/smartcontractkit/external-adapters-js",
-    "type": "git"
-  },
-  "license": "MIT",
-  "scripts": {
-    "clean": "rm -rf dist && rm -f tsconfig.tsbuildinfo",
-    "prepack": "yarn build",
-    "build": "tsc -b",
-    "server": "node -e 'require(\"./index.js\").server()'",
-    "server:dist": "node -e 'require(\"./dist/index.js\").server()'",
-    "start": "yarn server:dist"
-  },
-  "dependencies": {
-    "@chainlink/ea-bootstrap": "workspace:*",
-    "@chainlink/example-source-adapter": "workspace:*",
-    "tslib": "^2.3.1"
-  },
-  "devDependencies": {
-    "@types/jest": "27.5.2",
-    "@types/node": "16.11.51",
-    "typescript": "5.0.4"
-  }
-}
