--- packages/examples/composite/schemas/env.json
@@ -1,21 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/example-composite-adapter.json",
-  "title": "Chainlink Composite External Adapter for Example Data Provider",
-  "required": ["ETHEREUM_RPC_URL"],
-  "type": "object",
-  "properties": {
-    "ETHEREUM_RPC_URL": {
-      "type": "string",
-      "format": "uri"
-    },
-    "OPTION": {
-      "default": true,
-      "type": "boolean"
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
