--- packages/examples/composite/src/index.ts
@@ -1,8 +1,0 @@
-import { expose } from '@chainlink/ea-bootstrap'
-import { makeExecute } from './adapter'
-import { makeConfig, NAME } from './config'
-
-const adapterContext = { name: NAME }
-
-const { server } = expose(adapterContext, makeExecute())
-export { NAME, makeExecute, makeConfig, server }
