--- packages/examples/source/README.md
@@ -1,3 +1,0 @@
-# Chainlink External Adapter for Example
-
-This README will be generated automatically when code is merged to `develop`. If you would like to generate a preview of the README, please run `yarn generate:readme <ADAPTER_NAME>`.
