--- packages/examples/source/schemas/env.json
@@ -1,18 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/example-source-adapter.json",
-  "title": "Example Source Adapter",
-  "description": "Example Source Adapter (this title and description were pulled from source/schemas/env.json)",
-  "required": ["API_KEY"],
-  "type": "object",
-  "properties": {
-    "API_KEY": {
-      "type": "string",
-      "description": "API key to use"
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
