--- packages/examples/source/src/config/symbols.json
@@ -1,1 +1,0 @@
-{}
