--- packages/examples/source/src/endpoint/index.ts
@@ -1,5 +1,0 @@
-import type { TInputParameters as ExampleInputParameters } from './example'
-
-export type TInputParameters = ExampleInputParameters
-
-export * as example from './example'
