--- packages/examples/source/src/types.ts
@@ -1,3 +1,0 @@
-export interface ExampleType {
-  example: { type: '' }
-}
