--- packages/examples/source/test/integration/README.md
@@ -1,3 +1,0 @@
-Tests that run the adapter and send an end to end request using mocked API endpoints.
-
-See `packages/source/curve/test/integration` and `packages/composite/synth-index/test/integration` for reference. The integration tests are used for the example input/outputs during README generation.
