+++ packages/scripts/src/new/lib.ts
@@ -1,79 +1,6 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
-import chalk from 'chalk'
-import * as path from 'path'
-import * as shell from 'shelljs'
 import { getWorkspaceAdapters, WorkspaceAdapter } from '../workspace'
-const { red, blue } = chalk
-const { log } = console
-
-const ADAPTER_TYPES = ['composite', 'source']
-
-interface Inputs {
-  type: string
-  n: string
-}
-
-function writeJson(data: any) {
-  const files = Object.keys(data)
-
-  // write to each file
-  files.forEach((f) => {
-    let contents = data[f]
-    if (f.includes('.json')) {
-      contents = JSON.stringify(contents, null, 2)
-    }
-    shell.ShellString(contents).to(f)
-  })
-
-  // resolve workspace and format modified files
-  shell.exec(`yarn && yarn prettier --write ${files.join(' ')}`)
-}
-
-function checks(): Inputs {
-  const type: string = process.argv[2]
-  if (!type) throw red.bold('Missing first argument: type')
-  if (!ADAPTER_TYPES.includes(type))
-    throw red.bold(`Type must be one of: ${ADAPTER_TYPES.join(', ')}`)
-
-  const n: string = process.argv[3]
-  if (!n) throw red.bold('Missing second argument: name')
-
-  // check if jq is installed (jq used later to modify json files)
-  const jq: string = shell.exec('command -v jq').toString()
-  if (!jq) throw red.bold('jq is not installed')
-
-  return { type, n }
-}
-
-function copyFiles(type: string, n: string) {
-  // copying files
-  shell.mkdir(`packages/${type}s/${n}`)
-  shell.cp('-R', `packages/examples/${type}/*`, `packages/${type}s/${n}`)
-
-  // editing adapter package.json
-  shell
-    .cat(`packages/${type}s/${n}/package.json`)
-    .exec(
-      `jq '.name = "@chainlink/${n}-adapter" | .description = "Chainlink ${n} adapter." | .keywords += ["${n}"]'`,
-    )
-    .exec(`tee packages/${type}s/${n}/package.json`)
-    .to(`packages/${type}s/${n}/package.json`)
-
-  // clear the CHANGELOG
-  shell.echo(`# @chainlink/${n}-adapter\n`).to(`packages/${type}s/${n}/CHANGELOG.md`)
-
-  // set the new adapter version to 0.0.0
-  shell.sed(
-    '-i',
-    '"version": ".+?(?=",)',
-    '"version": "0.0.0',
-    `packages/${type}s/${n}/package.json`,
-  )
-
-  // changing README to use the adapter name instead of example
-  const nCap: string = n[0].toUpperCase() + n.slice(1)
-  shell.sed('-i', 'Example', nCap, `packages/${type}s/${n}/README.md`)
-}
+import path from 'path'
+import * as shell from 'shelljs'
 
 function tsconfGenerate(
   packages: WorkspaceAdapter[],
@@ -90,7 +17,7 @@ function tsconfGenerate(
   })
 }
 
-async function generate() {
+export async function generate() {
   let writeData = {} // data struct for writing
 
   // pull latest workspace data after files have been generated
@@ -100,12 +27,14 @@ async function generate() {
 
   // add to packages/tsconfig.json
   const tsconfigPath = 'packages/tsconfig.json'
+  /* eslint-disable @typescript-eslint/no-var-requires */
   const tsconfig = JSON.parse(JSON.stringify(require(path.relative(__dirname, tsconfigPath))))
   tsconfig.references = tsconfGenerate(currentWorkspace, tsconfigPath, 1)
   writeData = { ...writeData, [tsconfigPath]: tsconfig }
 
   // add to packages/tsconfig.test.json
   const tsconfigTestPath = 'packages/tsconfig.test.json'
+  /* eslint-disable @typescript-eslint/no-var-requires */
   const tsconfigTest = JSON.parse(
     JSON.stringify(require(path.relative(__dirname, tsconfigTestPath))),
   )
@@ -115,16 +44,15 @@ async function generate() {
   return writeData
 }
 
-export async function main(): Promise<void> {
-  log(blue.bold('Running input checks'))
-  const inputs: Inputs = checks()
+export function writeJson(data: any) {
+  const files = Object.keys(data)
 
-  log(blue.bold(`Copying example ${inputs.type} adapter to ${inputs.type}/${inputs.n}`))
-  copyFiles(inputs.type, inputs.n)
-
-  log(blue.bold('Regenerating tsconfig'))
-  const data = await generate()
-
-  log(blue.bold('Resolving workspace and running prettier'))
-  writeJson(data)
+  // write to each file
+  files.forEach((f) => {
+    let contents = data[f]
+    if (f.includes('.json')) {
+      contents = JSON.stringify(contents, null, 2)
+    }
+    shell.ShellString(contents).to(f)
+  })
 }
