+++ packages/tsconfig.test.json
@@ -108,12 +108,6 @@
       "path": "./core/test-helpers/tsconfig.test.json"
     },
     {
-      "path": "./examples/composite/tsconfig.test.json"
-    },
-    {
-      "path": "./examples/source/tsconfig.test.json"
-    },
-    {
       "path": "./non-deployable/reduce/tsconfig.test.json"
     },
     {
